refactor: extract OAuthUserProfile and JwtProperties to separate files
  - Extract OAuthUserProfile data class from OAuth2UserService to improve separation of concerns
  - Move JwtProperties configuration class to its own file
  - Simplify OAuth2UserService by delegating user creation logic to extracted profile class
  - Improve code organization and maintainability
