+++ src/main/kotlin/com/flownews/config/security/JwtProperties.kt
@@ -1,0 +1,9 @@
+package com.flownews.config.security
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "jwt")
+data class JwtProperties(
+    val secret: String?,
+    val expiration: Long,
+)
