+++ src/main/resources/application.yml
@@ -19,7 +19,7 @@ spring:
     static-path-pattern: /content/**
 
   cors:
-    allowed-origins: ${ALLOWED_ORIGINS:http://localhost:3000,https://sijeom.kr,http://121.189.148.34:5173,http://121.189.148.13:3000}
+    allowed-origins: ${ALLOWED_ORIGINS:http://localhost:3000,https://sijeom.kr}
 
   security:
     oauth2:
